requirements.txt
torch==1.8.1
transformers==4.5.1
numpy==1.20.2
faiss-gpu==1.7.0
shim